extract parameter details to dedicated file
